OE-7515 Fixed inability to edit 'Other' allergies in examination
The javascript controller didn't correctly check for the 'Other' option when removing duplicates from the list to add from.
I've also removed a few magic number 17s from the code. There is now a variable in both the js controller and the AllergyEntry classes. The latter also has an added function isOther().