OE-7683 - Allow prescribed medication end-date to be same as start-date
The start-dates were strings of the form "yyyy-mm-dd hh:mm:ss", but the
end-dates were of the form "yyyy-mm-dd". This caused the afterValidate()
method of HistoryMedicationsEntry to give a false error message and
prevent creation of the examination if end-date==start-date.

New version of afterValidate() method uses DateTime to remove the
hh:mm:ss and then compares.