OE-7308 Set onbeforeunload to null when patient ticketing is being run
Unfortunately there doesn't seem to be a way (outside of a much more extensive rework) to prevent the user from seeing unnecessary 'leaving page' prompts without disabling the prompt altogether. This takes effect whether the prompt is erroneous or not. The ticketing data seems safe but any events can lose data if exited in the wrong way with patient ticketing open.