OE-7897 - Allow patient hotlist to remain on screen when opened by hover
The patient hotlist was opening as expected when hovering the cursor
over the hotlist button, but it would disappear when attempting to move
the cursor onto the hotlist itself. This commit fixes that problem by
altering OpenEyes.UI.NavBtnPopup.js and using that NavBtnPopup prototype
instead of OpenEyes.UI.NavBtnSidebar.HotList.js. The actual features of
the hotlist (such as the datepicker and comment button) do not work
after this commit - those features need to be moved across from
OpenEyes.UI.NavBtnSidebar.js and OpenEyes.UI.NavBtnSidebar.HotList.js