OE-7521 - Make patient popups 'mouseenter' event overrule latched status
A change to the behaviour of the clinical patient popup headers was
requested. Previously, if one of them was latched, the 'mouseenter'
event listener was switched off for all of them.

This commit changes the behaviour so that if patient popup A is
latched, hovering the mouse over patient popup B will unlatch and close
patient popup A and show patient popup B (but will not latch popup B).