OE-7897 - Allow patient hotlist to remain on screen when opened by hover (#2385)
* OE-7897 - Allow patient hotlist to remain on screen when opened by hover

The patient hotlist was opening as expected when hovering the cursor
over the hotlist button, but it would disappear when attempting to move
the cursor onto the hotlist itself. This commit fixes that problem by
altering OpenEyes.UI.NavBtnPopup.js and using that NavBtnPopup prototype
instead of OpenEyes.UI.NavBtnSidebar.HotList.js. The actual features of
the hotlist (such as the datepicker and comment button) do not work
after this commit - those features need to be moved across from
OpenEyes.UI.NavBtnSidebar.js and OpenEyes.UI.NavBtnSidebar.HotList.js

* OE-7897 Move HotList's javascript from the sidebar to its own object

* OE-7897 - Remove unnecessary conditional call

* OE-7897 Change NavBtnPopUp to be expendable .Extend NavBtnPopUp from NavBtnPopUp.HotList

* OE-7897 Fix codefactor issues

* OE-7897 Fix codefactor issues

* OE-7897 Remove console.log , format lines

* OE-7897 Delete unused variable

* OE-7897 - Make hotlist sidebar fixed when window is wide enough

* OE-7877 - remove call to console.log

* OE-7897 - Remove unused method parameters

* OE-7897 - Make popups check if they are fixable on the page before fix