OE-7922 - Fix x-axes of graphs in OEScape to show correct date range
If empty graphs are present, plotly.js automatically assigns them an
x-axis range of -1 to 6. This was causing the earliest date to get
calculated incorrectly as 1970-01-01 (midnight UTC/GMT) when -1 was
converted to a date.

This commit fixes this problem by checking that the minimum and maximum
date calculations ignore axes that consist of numbers (using isNaN() to
check).