OE-8045 - Stop unopened child elements of list from collapsing group
The function that loads the active elements of the list view relied on
invoking the click event for the group headers. The fact that the list
headers did not respond to being clicked on when the group was already
expanded was necessary because the previous logic relied on this.

This commit alters the logic so that when loading the 'opened'
(highlighted) list element, the javascript loadElement() checks whether
the group is collapsed before invoking the click event on the header.