OE-8566 Previously when deleting a disorder the PCR risk was not notified , but even if you do the fields would not be updated as the PCR risk change event is checking for glaucoma disorders that's in the table and then change the fields accordingly and does nothing if there is no current glaucoma diagnoses.
When a glaucoma diagnosis is removed additional functionality was provided to handle that case. We only change Glaucoma field to NK if the diagnosis that was removed is glaucoma.

Also there could be a case where left eye has glaucoma and the you add an empty glaucoma diagnosis and decided to remove it , in that case it would check all the current glaucoma diagnoses and their selected eyes , but in this case it would be only the left one and if it manually was set to NO it will reset to Not Known to prevent this we also send as a parameter the eyes that were selected for the deleted diagnosis to only change the value to NK

IF

There is no more glaucoma diagnosis selected for that EYE and the deleted glaucoma diagnosis had that EYE selected