Fixed issues identified by Zhaolian
Zhaolian noticed that Canned report cataract 1-4 plots download files without '.csv' , this was due to the naming given to these files in the javascript did not include a '.csv' extension, I have added this.

Cannot download "NOD Audit" csv file, This was because the NOD report was left out of the options for the Download CSV button, I have added it with similar formatting to the other cataract reports.