OE-9254 Fixed OEScape crash.
In commit 0a8187499025f957a4bf0da65d814b13597410c8 the `run_oescape`
function was given an argument in the Visual Actuity History widget of
which the Medical Retinal History widget is a child. the MRH method does
not specify any arguments, php5 is allowing this overload but php7 is
not.