Feature/oe 9431 (#5439)
* Added export button to view screen for correspondence events.
Clicking the export button will attempt to export the PDF of the correspondence event to the specified portal using a SOAP request. The WSDL URL is set in the config parameters.
NOTE: The JS handler is currently assuming the response string is a SOAP response from CAVPMS. This will be made more generic in a future commit.
Added composer dependency for ext-soap (the PHP SOAP extension).

* Handling of SOAP response is now performed by the PHP application rather than Javascript.
Location URL is now specified as the correspondence export URL with any potential ?WSDL argument removed.
Host will now be verified automatically.
Fixed issues where alert dialogs on success or error would not open up, and would have no content and would enable buttons before the user has dismissed the alert.
Added correspondence_export_url to core params (currently defaults to empty string). This should be overridden in local confiig.

* Moved most of the export code from the controller to the export function within the model.
Removed underscore in key string.
Now returning the result of the export action as a JSON object.

* Restructured export function to allow usage of mock objects for unit testing purposes.
When running unit tests, the tests now run from the same working directory as the main application (when running PHPUnit from different paths this was causing issues when including composer packages in code such as FPDI).
Added Exportable interface which will be used to define elements that can be exported externally using web services. Current function stubs are getExportUrl and export.
Added test cases for ElementLetter exporting, which utilises a mock interface for the SOAP client.
Added key strings to letter fixtures.
Exceptions are now thrown from the export function if a specified web service type is unsupported or the export URL is not specified.

* Added empty correspondence_export_url parameter to common.sample.php.
This is to simplify setting the URL on new environments where the sample file is used for local config.

* Fixed potential issue where patient_risk_assignment view may not exist when the migration is run.

* Fixed some syntax errors in BaseEventTypeController and ElementLetter.
Added logging whenever a user-defined exception is thrown when exporting a correspondence event to an external service.

* [CodeFactor] Apply fixes

* Fixed missing event bind for exporting correspondence events to an external service.

* Prevented potential submit action when attempting to export a correspondence event to an external portal.

* Added 404 Not Found exception when attempting to export a correspondence event that doesn't exist.
Disabled autoprint injection on PDFs generated purely for external export.
Removed JSON parsing from AJAX handler for export button as the response is auto-parsed from the HTTP MIME type.

* Fixed null event issue when attempting to export a correspondence event.

* Set the controller's event to the letter's event when exporting correspondence. This will ensure Puppeteer can correctly generate the PDF file.

* Removed unnecessary logging.
Now deleting the generated PDF file after exporting it.

* Removed explicit location setting as we want to use the WSDL settings in a production scenario.
Restructured SOAP body.

* Added preg_replace to filter out the Hosnum/H tokens when a hospital number has been saved manually.

* Fixed failing unit tests for ElementLetter relating to export feature.

* Now replacing the H: prefix with U to match the Wales CRN pattern.

* Added padding to hospital number when exporting correspondence.

* Removed destructor for PuppeteerBrowser component as PuPHPeteer will automatically close the browser instance at the end of a request.
Added a base 64 conversion for the PDF file content to ensure it is transmitted correctly.
Now checking for the existence of the DocId property to determine success or failure.

* Added default correspondence URL to common.sample.php.

Co-authored-by: codefactor-io <support@codefactor.io>